refactor enable tha
